requirements.txt
langchain==0.1.0
langchain-core==0.1.23
langchain-community==0.0.10
langchainhub==0.1.0
langchain-dashscope==0.1.8

dashscope==1.18.0

python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
pandas==2.0.0
openpyxl==3.1.2  # Excel支持
pillow==10.0.0   # 图片处理

jupyter==1.0.0
ipykernel==6.0.0
pytest==7.4.0
black==23.9.1

google-search-results==2.4.2